backend/requirements.txt
Django>=5.1,<6.0
djangorestframework>=3.15,<4.0
django-cors-headers>=4.6,<5.0
channels>=4.1,<5.0
celery>=5.4,<6.0
redis>=5.2,<6.0
psycopg[binary]>=3.2,<4.0
python-dotenv>=1.0,<2.0
argon2-cffi>=23.1,<24.0
cryptography>=44.0,<45.0
